Removing Scala 3 http4s package object.
